Ising Cellular Automata
This file contains a Python script that performs two functions: First, it graphs the average magnetic moment of an electron network in a state of equilibrium at 9 different temperatures.  Second, it generates animations showing spin network evolution for 3 of the temperatures. 
It does this by randomly generating a 50x50 grid of spin up and spin down electrons.  A random electron's spin is then flipped; if the energy decreases, this becomes the new configuration.  If energy does not decrease, probability determines whether the configuration is updated.  This process repeats 600000 times before convergence is assumed and the magnetic moment is calculated.  For the graph, 5 independent magnetic moments are calculated for each temperature, and the largest is graphed.