Kinetic Gas Theory: 400 Particles in a Box
This code creates a movie of a two-dimensional box of 400 particles, as well as graphing the probability distributions of the velocities (which is an Maxwell-Boltzmann distribution) and kinetic energies of the gasses as histograms.  These histograms are fitted with analytical solutions to solve for temperature, which is written to a text file.